+++ src/YearZeroRoll.js
@@ -11,7 +11,7 @@ import { GameTypeError } from './errors.js';
  * Custom Roll class for Year Zero games.
  * @extends {Roll} The Foundry Roll class
  */
-export class YearZeroRoll extends Roll {
+export default class YearZeroRoll extends Roll {
   /**
    * @param {string} formula  The string formula to parse
    * @param {Object} data     The data object against which to parse attributes within the formula
@@ -266,10 +266,10 @@ export class YearZeroRoll extends Roll {
    * @param {string}         [title]      The name of the roll
    * @param {GameTypeString} [yzGame]     The game used
    * @param {number}         [maxPush=1]  The maximum number of pushes
-   * @param {boolean}        [push=false] Whether to add a push modifier to the roll
-   * @override
+   * @returns {YearZeroRoll}
+   * @static
    */
-  static createFromDiceQuantities(dice = {}, { title, yzGame = null, maxPush = 1, push = false } = {}) {
+  static forge(dice = {}, { title, yzGame = null, maxPush = 1, push = false } = {}) {
     // Checks the game.
     yzGame = yzGame ?? CONFIG.YZUR?.game;
     if (!YearZeroRollManager.GAMES.includes(yzGame)) throw new GameTypeError(yzGame);
@@ -295,6 +295,14 @@ export class YearZeroRoll extends Roll {
     const roll = new YearZeroRoll(formula, { name: title, game: yzGame, maxPush });
     if (CONFIG.debug.dice) console.log(roll);
     return roll;
+  }
+
+  /** @deprecated */
+  // eslint-disable-next-line no-unused-vars
+  static createFromDiceQuantities(dice = {}, { title, yzGame = null, maxPush = 1, push = false } = {}) {
+    // eslint-disable-next-line max-len
+    console.warn('YZUR | createFromDiceQuantities() is deprecated and will be removed in a future release. Use forge()instead.');
+    return YearZeroRoll.forge(dice, { title, yzGame, maxPush });
   }
 
   /* -------------------------------------------- */
@@ -400,7 +408,7 @@ export class YearZeroRoll extends Roll {
     }
     // If the DieTerm doesn't exist, creates it.
     else {
-      const cls = YZUR.DICE.DIE_TYPES[type];
+      const cls = CONFIG.YZUR.DICE.DIE_TYPES[type];
       term = new cls({
         number: qty,
         faces: range,
@@ -534,9 +542,9 @@ export class YearZeroRoll extends Roll {
   async modify(mod = 0) {
     if (!mod) return this;
 
-    // TWILIGHT 2000
+    // TWILIGHT 2000 & BLADE RUNNER
     // --------------------------------------------
-    if (this.game === 't2k') {
+    else if (this.game === 't2k' || this.game === 'br') {
       const diceMap = [null, 6, 8, 10, 12, Infinity];
       const typesMap = ['d', 'd', 'c', 'b', 'a', 'a'];
       const refactorRange = (range, n) => diceMap[diceMap.indexOf(range) + n];
@@ -545,38 +553,60 @@ export class YearZeroRoll extends Roll {
       const _terms = this.getTerms('base');
       const dice = _terms.flatMap(t => new Array(t.number).fill(t.faces));
 
-      // 1 — Modifies the dice ranges.
-      while (mod !== 0) {
-        let i;
-        // 1.1.1 — A positive modifier increases the lowest term.
+      // BLADE RUNNER
+      if (this.game === 'br') {
+        // Gets the lowest term.
+        const lowest = Math.min(...dice);
+
+        // A positive modifier means advantage.
+        // An advantage adds a third base die, same value as lowest.
         if (mod > 0) {
-          i = dice.indexOf(Math.min(...dice));
-          dice[i] = refactorRange(dice[i], 1);
-          mod--;
+          dice.push(lowest);
         }
-        // 1.1.2 — A negative modifier decreases the highest term.
-        else {
-          i = dice.indexOf(Math.max(...dice));
-          dice[i] = refactorRange(dice[i], -1);
-          mod++;
+        // A negative modifier means disadvantage.
+        // A disadvantage removes the lowest die.
+        else if (mod < 0) {
+          const i = dice.indexOf(lowest);
+          dice.splice(i, 1);
         }
-        // 1.2 — Readjusts term faces.
-        if (dice[i] === Infinity) {
-          dice[i] = refactorRange(dice[i], -1);
-          if (dice.length < 2) {
-            dice.push(diceMap[1]);
-          }
-        }
-        else if (dice[i] === null) {
-          if (dice.length > 1) {
-            dice.splice(i, 1);
-          }
-          else {
+        mod = 0;
+      }
+
+      // TWILIGHT 2000
+      else {
+        // 1 — Modifies the dice ranges.
+        while (mod !== 0) {
+          let i;
+          // 1.1.1 — A positive modifier increases the lowest term.
+          if (mod > 0) {
+            i = dice.indexOf(Math.min(...dice));
             dice[i] = refactorRange(dice[i], 1);
+            mod--;
           }
-        }
-        else if (dice[i] === undefined) {
-          throw new Error(`YZUR | YearZeroRoll#modify<T2K> | dice[${i}] is out of bounds (mod: ${mod})`);
+          // 1.1.2 — A negative modifier decreases the highest term.
+          else {
+            i = dice.indexOf(Math.max(...dice));
+            dice[i] = refactorRange(dice[i], -1);
+            mod++;
+          }
+          // 1.2 — Readjusts term faces.
+          if (dice[i] === Infinity) {
+            dice[i] = refactorRange(dice[i], -1);
+            if (dice.length < 2) {
+              dice.push(diceMap[1]);
+            }
+          }
+          else if (dice[i] === null) {
+            if (dice.length > 1) {
+              dice.splice(i, 1);
+            }
+            else {
+              dice[i] = refactorRange(dice[i], 1);
+            }
+          }
+          else if (dice[i] === undefined) {
+            throw new Error(`YZUR | YearZeroRoll#modify<T2K> | dice[${i}] is out of bounds (mod: ${mod})`);
+          }
         }
       }
       // 2 — Filters out all the base terms.
@@ -614,7 +644,10 @@ export class YearZeroRoll extends Roll {
     // --------------------------------------------
     else {
       const skill = this.count('skill');
-      if (mod < 0) mod = Math.max(-skill + 1, mod); // Minimum of 1 skill die
+      if (mod < 0) {
+        // Minimum of 1 skill die.
+        mod = Math.max(-skill + 1, mod);
+      }
       await this.addDice(mod, 'skill');
     }
 
@@ -628,146 +661,6 @@ export class YearZeroRoll extends Roll {
     }
 
     return this;
-  }
-
-  /**
-   * @deprecated This is the old modify method.
-   * TODO remove
-   */
-  _modify(mod) {
-    // eslint-disable-next-line max-len
-    console.warn('YZUR | You are using the old _modify() method, which is deprecated and will be removed in a future release.');
-    // Exits early if no modifier.
-    if (!mod) return this.duplicate();
-
-    // Gets the dice quantities.
-    const dice = this.getDiceQuantities();
-
-    let occurenceNb = 0;
-    while (mod !== 0) {
-      // Failsafe – Watches the number of occurences to avoid infinite loops.
-      occurenceNb++;
-      if (occurenceNb >= 100) throw new RangeError(`${this.constructor.name} | Infinite modify loop!`);
-
-      // TWILIGHT 2000
-      if (this.game === 't2k') {
-        const dieTypes = ['d', 'c', 'b', 'a'];
-
-        // Creates a dice pool array and finds the total quantities of each die.
-        const pool = Object.entries(dice).reduce((arr, [k, v]) => {
-          if (dieTypes.includes(k)) {
-            for (; v > 0; v--) arr.push(k);
-          }
-          return arr;
-        }, []);
-        const n = pool.length;
-
-        // Early exits.
-        if (mod > 0) {
-          if (n > 2) break;
-          if (pool.filter(t => t === 'a').length >= 2) break;
-        }
-        else if (n === 0) {
-          dice.d = 1;
-          break;
-        }
-        else if (n === 1 && pool.includes('d')) {
-          break;
-        }
-
-        // Initializes null dice.
-        for (const type of dieTypes) if (!dice[type]) dice[type] = 0;
-
-        // Gets the die to modify.
-        // For a positive modifier, we take the lowest die.
-        // For a negative modifier, we take the highest one.
-        const die = pool.reduce((a, b) => {
-          if (mod > 0) {
-            if (b === 'a') return a;
-            return a > b ? a : b;
-          }
-          return a < b ? a : b;
-        }, undefined);
-
-        // Modifies the range.
-        const currentRangeIndex = dieTypes.indexOf(die);
-        let newDie;
-        if (currentRangeIndex >= 0) {
-          const maxRangeIndex = dieTypes.length - 1;
-          const rangeIndex = currentRangeIndex + mod;
-          const newRangeIndex = Math.clamped(rangeIndex, 0, maxRangeIndex);
-          newDie = dieTypes[newRangeIndex];
-          mod -= (newRangeIndex - currentRangeIndex);
-          dice[die]--;
-          dice[newDie]++;
-        }
-
-        // Positive excess mod means adding an extra die.
-        // Note: the pool can only have a maximum of 2 dice.
-        if (mod > 0) {
-          if (n < 2) {
-            const ex = Math.min(dieTypes.length, mod);
-            dice[dieTypes[ex - 1]]++;
-            if (mod > ex) mod -= ex;
-            else break;
-          }
-          else {
-            const diceBelowMaxRange = Object.entries(dice).filter(([k, v]) => v > 0 && k > 'a').length;
-            if (diceBelowMaxRange <= 0) break;
-          }
-        }
-        // Negative excess mod means removing the die and decreasing another one.
-        // Note: The pool has always 1 die.
-        else if (mod < 0 && n > 1) {
-          dice[newDie]--;
-          // We add 1 because we removed one die (which is 1 step).
-          mod++;
-        }
-      }
-      // MUTANT YEAR ZERO & FORBIDDEN LANDS
-      else if (this.game === 'myz' || this.game === 'fbl') {
-        // Balances skill & neg dice.
-        if (dice.skill > 0 && dice.neg > 0) {
-          while (dice.skill > 0 && dice.neg > 0) {
-            dice.skill--;
-            dice.neg--;
-          }
-        }
-        if (!dice.skill) dice.skill = 0;
-        const neg = Math.max(0, -mod - dice.skill);
-        dice.skill = Math.max(0, dice.skill + mod);
-        if (neg > 0) {
-          if (!dice.neg) dice.neg = 0;
-          dice.neg += neg;
-        }
-        mod = 0;
-      }
-      // ALL OTHER GAMES
-      else {
-        if (!dice.skill) dice.skill = 0;
-        dice.skill = Math.max(1, dice.skill + mod);
-        mod = 0;
-      }
-    }
-
-    // Builds the new roll instance.
-    const _roll = YearZeroRoll.createFromDiceQuantities(dice);
-    const _data = _roll.toJSON();
-    const data = this.toJSON();
-    // Keeps options.
-    for (const t of data.terms) {
-      if (!foundry.utils.isObjectEmpty(t.options)) {
-        for (const _t of _data.terms) {
-          if (_t.type === t.type) {
-            _t.options = t.options;
-            break;
-          }
-        }
-      }
-    }
-    data.terms = _data.terms;
-    data.formula = _data.formula;
-    return this.constructor.fromData(data);
   }
 
   /* -------------------------------------------- */
@@ -788,63 +681,6 @@ export class YearZeroRoll extends Roll {
         return at - bt;
       });
     // <==
-    // TODO clean this commented-out code at next Foundry version.
-    // const parts = this.dice.map(d => {
-    //   const cls = d.constructor;
-    //   return {
-    //     formula: d.formula,
-    //     total: d.total,
-    //     faces: d.faces,
-    //     // ==>
-    //     // // flavor: d.options.flavor,
-    //     flavor: d.options.flavor || (
-    //       CONFIG.YZUR?.DICE?.localizeDieTypes
-    //         ? game.i18n.localize(`YZUR.DIETYPES.${cls.name}`)
-    //         : null
-    //     ),
-    //     number: d.number,
-    //     // // rolls: d.results.map(r => {
-    //     rolls: d.results.map((r, i) => {
-    //       // <==
-    //       const hasSuccess = r.success !== undefined;
-    //       const hasFailure = r.failure !== undefined;
-    //       // ==>
-    //       // // const isMax = r.result === d.faces;
-    //       // // const isMin = r.result === 1;
-    //       let isMax = false, isMin = false;
-    //       if (d.type === 'neg') {
-    //         isMax = false;
-    //         isMin = r.result === 6;
-    //       }
-    //       else {
-    //         isMax = r.result === d.faces || r.count >= 1;
-    //         isMin = r.result === 1 && d.type !== 'skill' && d.type !== 'loc';
-    //       }
-    //       // <==
-    //       return {
-    //         result: cls.getResultLabel(r.result),
-    //         // ==>
-    //         row: r.indexPush,
-    //         col: r.indexResult,
-    //         // <==
-    //         classes: [
-    //           cls.name.toLowerCase(),
-    //           'd' + d.faces,
-    //           r.success ? 'success' : null,
-    //           r.failure ? 'failure' : null,
-    //           r.rerolled ? 'rerolled' : null,
-    //           r.exploded ? 'exploded' : null,
-    //           r.discarded ? 'discarded' : null,
-    //           // ==>
-    //           r.pushed ? 'pushed' : null,
-    //           // <==
-    //           !(hasSuccess || hasFailure) && isMin ? 'min' : null,
-    //           !(hasSuccess || hasFailure) && isMax ? 'max' : null,
-    //         ].filter(c => c).join(' '),
-    //       };
-    //     }),
-    //   };
-    // });
     // START MODIFIED PART ==>
     if (this.pushed) {
       // Converts "parts.rolls" into a matrix.
@@ -935,6 +771,7 @@ export class YearZeroRoll extends Roll {
       infos: isPrivate ? null : await this.getRollInfos(chatOptions.infosTemplate),
       pushable: isPrivate ? false : this.pushable,
       options: chatOptions,
+      isPrivate,
       roll: this,
     };
 
@@ -982,22 +819,6 @@ export class YearZeroRoll extends Roll {
   /* -------------------------------------------- */
   /*  JSON                                        */
   /* -------------------------------------------- */
-
-  // TODO clean
-  // /** @override */
-  // static fromData(data) {
-  //   const roll = super.fromData(data);
-  //   roll.data = data.data ?? {};
-  //   return roll;
-  // }
-
-  // /** @override */
-  // toJSON() {
-  //   return {
-  //     ...super.toJSON(),
-  //     data: this.data,
-  //   };
-  // }
 
   /**
    * Creates a deep clone copy of the roll.